Improved sRGB support.
The following operations are now always performed in linear color space:
* Scaling.
* Pre-multiplication of alpha.
* Conversion to grayscale.

This will produce more accurate results, since the image's true color is in
linear space, while sRGB is the storage format. Images now store the color
space they are represented in to properly manage this, and you can now
freely convert between color spaces with Image::changeColorSpace().

Textures are now provided the color space in the constructor/initialize().
When the image color space doesn't match, it is converted.

Disabled special quality setting for sRGB when converting ASTC textures.
While this is done for the astcenc tool, it is incredibly slow.

Since this is a breaking API change, the version has been incremented to
2.0.0.