Perform release builds immediately after debug.
This ensures even further permutation testing, and is a model for how the
other projects will do CI. It will pretty much be required for the other
projects for managing Windows builds, and also avoids too many parallel
builds when more configurations are used. It does slightly reduce the
parallelization for this project, but given that the Windows build is the
bottleneck it shouldn't matter.