Integrated the bc7enc library.
This is a significantly faster (think seconds instead of 30 minutes)
bc7 compressor compared to the NVidia Texture Tools implementation. It's
not complete, implementing only block modes 1 and 6, but in most
situations are consistent.

For lowest through normal quality levels, bc7enc16 is the only compressor
used. For high quality level, it will only use the NVidia compressor if the
alpha is deemed complex. (difference between lowest and highest alpha value
is >= 0.2) For highest quality level, the NVidia compressor is always used
to take advantage of all 8 block modes.

Updated the submodules for NVidia Texture Tools and PVRTexToolLib and
incremented version to 2.1.0.

Resolves #3