Added Image::save() function.
Saves the image to a file with a format based on the extension. This is
intended to be used for the most common use cases, such as debugging or
outputting the result of a transformation as a viewable image rather than
a texture. The default values are used for settings such as compression.
The raw image data should be extracted and passed to a dedicated image
library if more fine-tuned control is required.

Incremented version number to 2.3.0.