Support tiled normal maps.
Added options to wrap along the X and/or Y axis when computing normal maps.
This will ensure continuous normals when tiling textures.

When using the API, the keepSign parameter has been replaced with a
NormalOptions bitmask enum. When using the CLI, "wrap", "wrapx", or "wrapy"
may be provided after the -n/--normalmap option to wrap along both axes,
the X axis, or Y axis, respectively.

Incremented version to 2.4.0.