More fixes for ASTC.
Updated to the latest master of astc-encoder. Use the higher level function
to populate the ASTC encoding block since a dummy image needs to be created
anyway. No longer initialize tables, which are initialized by the context.

Thread data is now created ahead of time before the threads start. This
avoids issues where the table initialization could be done simultaneously
across threads and cause data corruption.

Build on Visual Studio 2019 for Windows. There seems to be an issue with
HDR encoding for 32-bit Windows in release builds on VS 2017, which causes
the HDR test to fail. The ASTC encoding library uses some extreme floating
point values, so this combined with some optimizations could be the cause
of the problem.