Fixed warning (treated as error) on Mac builds.
For some reason this passed with the normal CI build, but not for the tools
build.

Updated astc-encoder. Unfortunately this can't be directly to the 2.2
release since it's missing a variable initialization, causing memory
errors.

Switched back to VS 2017 for Windows build. Most likely the failure was
caused by the missing initialized variable, which only showed up on other
platforms (like Mac) after other modifications changed the memory layout.