Switched encoders for S3TC formats.
Dropped NVidia Texture Tools since the encoders are quite old and
unmaintained. Since the repository is now archived, it's not expected to
change.

Use rgbxc for BC1-5 formats, except for the following cases it doesn't
support:
* BC1_RGBA uses squish for blocks with alpha. (would use Compressionator
  except it has alpha support disabled internally despite having the
  interface for it)
* BC2 populates the alpha block manually, using rgbxc for the color block.
* BC4/5 signed uses Compressonator.

When ISPC is available, ispc_texcomp is used for unsigned BC6H and bc7e
for BC7. Compressonator is used for signed BC6H and fallback for unsigned
BC6H as a fallback when ISPC isn't available. An updated version of bc7enc
is used when ISPC isn't available.

Incremented version to 2.5.0.