Avoid linking to PVRTexLib when PVRTC is disabled.
Only link to PVRTexLib if CUTTLEFISH_BUILD_PVRTC is set. This avoids
situations where the include dirs and library variables may be set with a
previous call to find. (e.g. toggling the option) Also avoid calling
find_package(PVRTexLib) in the tool for the Mac rpath fixup unless PVRTC
is enabled.