Use public interface for astcenc.
Was hesitant to use this before due to having to construct the image and
the fact that it required resetting the context. In practice this doesn't
seem to be a performance limitation.

Added a cache for ASTC contexts. This avoids re-computing the lookup tables
when possible, especially when multiple mipmaps or faces are compressed.

Switched to a fork of astc-encoder that includes a fix preventing
uninitialized memory access for invalid block modes. Adjusted lowest preset
to work around issue that this fixed introduced where enough block modes
being disabled prevents blocks from passing the error threshold.