Added CUTTLEFISH_STATIC_RUNTIME CMake option.
When set to ON, the static runtime library will be used on Windows.

Set the minimum required CMake version to 3.15 on Windows to have access
to set(CMAKE_MSVC_RUNTIME_LIBRARY. The minimum on other platforms is now
3.10 to prevent warnings on newer versions, as future versions will remove
support for compatibility with minimum versions < 3.9.

Set CMake policy for versions >= 3.27 to avoid warning when using
find_package() with PVRTexLib. The usage was already compliant with new
behavior.

Create separate "tool" and "full" packages for Windows in CI. "tool"
packages provide only the tool with static linking and runtime, while
"full" is same as the previous package with the addition of the debug
library.