Better handle signed and integer output types.
When running the CLI, if the original input image is a UNorm type and the
output texture type is SNorm, the values will be re-mapped from [0, 1] to
[-1, 1] before conversion. In the case of Int and UInt outputs, the value
will be converted to the full integer range.

Conversion to EAC signed formats now take into account the fact that it
assumes the inputs are in the range [0, 1].

Worked around bugs in FreeImage conversions:
* RGB5 and RGB565 to RGBF handles the values as a 24 BPP image.
* Grayscale values converted to double don't normalize the values.

Updated astc-encoder to 5.3.0. Updated ISPC and GTest versions sued by CI.
Incremented version to 2.9.0.