Added support to provide custom mips to the CLI tool.
Can pass custom mips via -M/--custom-mip. The level and file are provided,
along with optionally the depth, face, and replacement mode. The default
replacement node is "continue" to use the replacement for following mip
levels. This argument may be provided multiple times to replace multiple
mips.

The mip replacements may also be put in a file, with each line containing
the arguments for --custom-mip, and passed via --custom-mip-list. This
allows for a large list of custom mips to be provided without exceeding the
limits for command-line length.

Resolves #27